Support sub-second times in physics

+++ physics/physics.go
@@ -19,7 +19,7 @@ type Vel struct {
 // AsPosition computes the position from the origin that the velocity affects
 // over the given dt.
 func (v Vel) AsPosition(dt time.Duration) Pos {
-	secs := float32(dt / time.Second)
+	secs := float32(dt) / float32(time.Second)
 	return Pos{glm.Vec3{v.X() * secs, v.Y() * secs, v.Z() * secs}}
 }
 
@@ -31,6 +31,6 @@ type Acc struct {
 // AsVelocity computes the velocity from the origin that the acceleration
 // affects over the given dt.
 func (a Acc) AsVelocity(dt time.Duration) Vel {
-	secs := float32(dt / time.Second)
+	secs := float32(dt) / float32(time.Second)
 	return Vel{glm.Vec3{a.X() * secs, a.Y() * secs, a.Z() * secs}}
 }
